requirements.txt
Pillow
imageio
imageio-ffmpeg